third_party/WebKit/LayoutTests/fast/dom/CSSStyleDeclaration/css-style-declaration-named-setter.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body style="margin: 0px">
<script>

shouldBeEqualToString("document.body.style.margin", "0px");
shouldBe("document.body.style.margin = 1", "1");
shouldBeEqualToString("document.body.style.margin", "1px");

shouldBe("document.body.style.__proto__.margin = 777", "777");
shouldBe("document.body.style.__proto__.margin", "777");

shouldBeEqualToString("document.body.style.margin", "1px");
shouldBe("document.body.style.margin = 2", "2");
shouldBeEqualToString("document.body.style.margin", "2px");

</script>
</body>
</html>
